wip: refactor with CRUD
