--- src/gas/controllers/MembersController.ts
@@ -1,7 +1,0 @@
-export default class MembersController {
-  // TODO: 直す
-  //@ts-ignore
-  join(parmas) {
-    return parmas;
-  }
-}
