+++ src/gas/shared/Utils.ts
@@ -1,6 +1,6 @@
 import ResponseInterface from "../types/ResponseInterface";
 
-export default class Util {
+export default class Utils {
   constructor() {}
 
   makeSuccess(response: ResponseInterface) {
@@ -11,5 +11,11 @@ export default class Util {
   makeError(response: ResponseInterface) {
     response.success = false;
     return response;
+  }
+
+  createOutput(response?: ResponseInterface) {
+    return (
+      response && ContentService.createTextOutput(JSON.stringify(response)).setMimeType(ContentService.MimeType.JSON)
+    );
   }
 }
